feat: Use blueprint 0.58 isInt flag for integer types
Upgrade @seamapi/blueprint to 0.58.0 and map number parameters and
properties with isInt to Python int, restoring the integer types that
were collapsed to float when the codegen moved off the raw OpenAPI
schemas. List items carry no isInt flag, so numbers inside lists remain
float.

+++ seam/routes/access_codes.py
@@ -231,8 +231,8 @@ class AccessCodes(AbstractAccessCodes):
         self,
         *,
         device_id: str,
-        max_code_length: Optional[float] = None,
-        min_code_length: Optional[float] = None,
+        max_code_length: Optional[int] = None,
+        min_code_length: Optional[int] = None,
         supported_code_lengths: Optional[List[float]] = None
     ) -> None:
         json_payload = {}
