+++ seam/routes/acs_users.py
@@ -102,7 +102,7 @@ class AcsUsers(AbstractAcsUsers):
         *,
         acs_system_id: Optional[str] = None,
         created_before: Optional[str] = None,
-        limit: Optional[float] = None,
+        limit: Optional[int] = None,
         page_cursor: Optional[str] = None,
         search: Optional[str] = None,
         user_identity_email_address: Optional[str] = None,
